Admin space

+++ server/src/models/author.ts
@@ -66,16 +66,57 @@ export class Author {
     if (!author_id) {
       throw new Error("Author id is required!");
     }
+
     try {
-      const query = `DELETE FROM authors WHERE id = $1`;
-      const result = await pool.query(query, [author_id]);
+      // Check if the author is linked to any books
+      const checkQuery = `SELECT COUNT(*) FROM book_authors WHERE author_id = $1`;
+      const checkResult = await pool.query(checkQuery, [author_id]);
+
+      if (parseInt(checkResult.rows[0].count) > 0) {
+        throw new Error("Cannot delete author: Author has associated books!");
+      }
+
+      // Delete the author
+      const deleteQuery = `DELETE FROM authors WHERE id = $1`;
+      const result = await pool.query(deleteQuery, [author_id]);
 
       if (result.rowCount === 0) {
         throw new Error("Author not found!");
       }
     } catch (error: any) {
-      console.error("Error deleting authors:", error.message);
+      console.error("Error deleting author:", error.message);
       throw new Error(error.message);
+    }
+  }
+
+  static async getBooksByAuthorId(author_id: string) {
+    if (!author_id) {
+      throw new Error("Author ID is required!");
+    }
+
+    try {
+      const query = `
+      SELECT
+        b.id AS book_id,
+        b.title,
+        b.description,
+        b.published_date,
+        b.pages,
+        b.price,
+        b.cover_image_url,
+        b.state
+      FROM books b
+      INNER JOIN book_authors ba ON ba.book_id = b.id
+      WHERE ba.author_id = $1
+      ORDER BY b.title ASC
+    `;
+
+      const result = await pool.query(query, [author_id]);
+
+      return result.rows;
+    } catch (error: any) {
+      console.error("Error retrieving books for author:", error.message);
+      throw new Error("Error fetching books for author: " + error.message);
     }
   }
 }
